SSCS-11092 splits out HearingsDurationMappingAdjournmentTest
