+++ src/test/java/uk/gov/hmcts/reform/sscs/helper/mapping/HearingsDurationMappingTest.java
@@ -6,12 +6,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
-import uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCaseNextHearingDurationType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCaseNextHearingDurationUnits;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Adjournment;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
 import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitCode;
@@ -25,205 +21,24 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.OtherParty;
 import uk.gov.hmcts.reform.sscs.ccd.domain.OverrideFields;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SchedulingAndListingFields;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.ccd.domain.YesNo;
 import uk.gov.hmcts.reform.sscs.reference.data.model.HearingDuration;
 import uk.gov.hmcts.reform.sscs.reference.data.service.HearingDurationsService;
 import uk.gov.hmcts.reform.sscs.service.holder.ReferenceDataServiceHolder;
 
 import java.util.List;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCaseNextHearingDurationUnits.MINUTES;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCaseNextHearingDurationUnits.SESSIONS;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class HearingsDurationMappingTest extends HearingsMappingBase {
-    public static final int DURATION_FACE_TO_FACE = 60;
-    public static final int DURATION_INTERPRETER = 75;
-    public static final int DURATION_PAPER = 40;
     @Mock
     private HearingDurationsService hearingDurations;
 
-    private SscsCaseData adjournCaseBuilder(
-        Integer adjournCaseDuration,
-        AdjournCaseNextHearingDurationUnits adjournCaseDurationUnits
-    ) {
-        return SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .adjournment(Adjournment.builder()
-                .nextHearingListingDurationType(AdjournCaseNextHearingDurationType.NON_STANDARD)
-                .nextHearingListingDuration(adjournCaseDuration)
-                .nextHearingListingDurationUnits(adjournCaseDurationUnits)
-                .adjournmentInProgress(YesNo.YES)
-                .build())
-            .appeal(Appeal.builder()
-                .hearingOptions(HearingOptions.builder().build())
-                .build())
-            .build();
-    }
-
-    private SscsCaseData wantsToAttendCaseBuilder(String wantsToAttend) {
-        return SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .appeal(Appeal.builder()
-                .hearingOptions(HearingOptions.builder()
-                    .wantsToAttend(wantsToAttend)
-                    .build())
-                .build())
-            .build();
-    }
-
     @Mock
     private ReferenceDataServiceHolder referenceDataServiceHolder;
-
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    @Nested
-    class AdjournmentFlagEnabled {
-
-        @BeforeEach
-        void setUp() {
-            given(referenceDataServiceHolder.isAdjournmentFlagEnabled()).willReturn(true);
-        }
-
-        @DisplayName("When a valid adjournCaseDuration and adjournCaseDurationUnits is given "
-            + "getHearingDuration returns the correct duration Parameterized Tests")
-        @ParameterizedTest
-        @MethodSource
-        void getHearingDuration(Integer adjournCaseDuration, AdjournCaseNextHearingDurationUnits adjournCaseDurationUnits, int expected) {
-            SscsCaseData caseData = adjournCaseBuilder(
-                adjournCaseDuration,
-                adjournCaseDurationUnits
-            );
-
-            Integer result = HearingsDurationMapping.getHearingDuration(caseData, referenceDataServiceHolder);
-
-            assertThat(result).isEqualTo(expected);
-        }
-
-        private Stream<Arguments> getHearingDuration() {
-            return Stream.of(
-              Arguments.of(120, MINUTES, 120),
-              Arguments.of(70, MINUTES, 70),
-              Arguments.of(1, SESSIONS, 165),
-              Arguments.of(2, SESSIONS, 330),
-              Arguments.of(3, SESSIONS, 495)
-            );
-        }
-
-        @DisplayName("When adjournment flag is enabled but getHearingDurationAdjournment returns null "
-            + "uses default hearing duration")
-        @Test
-        void getHearingDurationAdjournedFeatureFlagEnabled() {
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-            given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE)).willReturn(null);
-
-            SscsCaseData caseData = SscsCaseData.builder()
-                .benefitCode(BENEFIT_CODE)
-                .issueCode(ISSUE_CODE)
-                .appeal(Appeal.builder()
-                    .hearingOptions(HearingOptions.builder()
-                        .wantsToAttend("Yes")
-                        .build())
-                    .build())
-                .build();
-
-            Integer durationAdjourned = HearingsDurationMapping.getHearingDurationAdjournment(
-                caseData, referenceDataServiceHolder);
-            assertThat(durationAdjourned).isNull();
-
-            Integer result = HearingsDurationMapping.getHearingDuration(
-                caseData,
-                referenceDataServiceHolder
-            );
-
-            assertThat(result).isEqualTo(HearingsDurationMapping.DURATION_DEFAULT);
-        }
-
-        @DisplayName("When a valid duration is given but adjournCaseDurationUnits is not provided "
-            + "getHearingDuration returns the default adjournment duration")
-        @Test
-        void getHearingDurationNullUnits() {
-            given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-                .willReturn(new HearingDuration(
-                    BenefitCode.PIP_NEW_CLAIM,
-                    Issue.DD,
-                    DURATION_FACE_TO_FACE,
-                    DURATION_INTERPRETER,
-                    DURATION_PAPER
-                ));
-
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-            SscsCaseData caseData = adjournCaseBuilder(2, null);
-
-            int result = HearingsDurationMapping.getHearingDuration(caseData, referenceDataServiceHolder);
-
-            assertThat(result).isEqualTo(DURATION_PAPER);
-        }
-
-        @DisplayName("When an invalid adjournCaseDuration and valid adjournCaseDurationUnits is given "
-            + "getHearingDuration a null pointer exception is thrown")
-        @ParameterizedTest
-        @MethodSource
-        void getHearingDurationFailure(Integer adjournCaseDuration, AdjournCaseNextHearingDurationUnits adjournCaseDurationUnits) {
-            SscsCaseData caseData = adjournCaseBuilder(adjournCaseDuration, adjournCaseDurationUnits);
-
-            assertThatThrownBy(() -> HearingsDurationMapping.getHearingDuration(caseData, referenceDataServiceHolder))
-                .isInstanceOf(NullPointerException.class);
-        }
-
-        private Stream<Arguments> getHearingDurationFailure() {
-            return Stream.of(
-              Arguments.of(null, SESSIONS),
-              Arguments.of(0, SESSIONS),
-              Arguments.of(0, MINUTES),
-              Arguments.of(null, MINUTES)
-            );
-        }
-
-
-        @DisplayName("When getAdjournCaseNextHearingListingDurationType is non standard and  "
-            + "nextHearingListingDuration is blank, getHearingDurationAdjournment returns null")
-        @Test
-        void getHearingDurationAdjournment_nextHearingListingDurationIsBlank() {
-            SscsCaseData caseData = adjournCaseBuilder(null, SESSIONS);
-
-            Integer result = HearingsDurationMapping.getHearingDurationAdjournment(caseData, referenceDataServiceHolder);
-
-            assertThat(result).isNull();
-        }
-
-        @DisplayName("When getAdjournCaseNextHearingListingDurationType is standard "
-            + "getHearingDurationAdjournment returns null")
-        @Test
-        void getHearingDurationAdjournment_nextHearingListingDurationTypeIsStandard() {
-            SscsCaseData caseData = adjournCaseBuilder(null, SESSIONS);
-            caseData.getAdjournment().setNextHearingListingDurationType(AdjournCaseNextHearingDurationType.STANDARD);
-
-            Integer result = HearingsDurationMapping.getHearingDurationAdjournment(caseData, referenceDataServiceHolder);
-
-            assertThat(result).isNull();
-        }
-    }
-
-    @DisplayName("When adjournment flag is disabled getHearingDurationAdjournment returns null")
-    @Test
-    void getHearingDurationAdjournedFeatureFlagDisabled() {
-        given(referenceDataServiceHolder.isAdjournmentFlagEnabled()).willReturn(false);
-
-        SscsCaseData caseData = adjournCaseBuilder(100, MINUTES);
-
-        Integer result = HearingsDurationMapping.getHearingDurationAdjournment(caseData, referenceDataServiceHolder);
-
-        assertThat(result).isNull();
-    }
 
     @TestInstance(TestInstance.Lifecycle.PER_CLASS)
     @Nested
@@ -231,6 +46,7 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
 
         @BeforeEach
         void setUp() {
+            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
             given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
                 .willReturn(new HearingDuration(
                 BenefitCode.PIP_NEW_CLAIM,
@@ -258,8 +74,6 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
                 any()
             ))
                 .willReturn(DURATION_INTERPRETER);
-
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
 
             SscsCaseData caseData = SscsCaseData.builder()
                 .benefitCode(BENEFIT_CODE)
@@ -291,9 +105,15 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
         @Test
         void getHearingDurationBenefitIssueCodesPaper() {
 
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-            SscsCaseData caseData = wantsToAttendCaseBuilder("No");
+            SscsCaseData caseData = SscsCaseData.builder()
+                .benefitCode(BENEFIT_CODE)
+                .issueCode(ISSUE_CODE)
+                .appeal(Appeal.builder()
+                    .hearingOptions(HearingOptions.builder()
+                        .wantsToAttend("No")
+                        .build())
+                    .build())
+                .build();
 
             Integer result = HearingsDurationMapping.getHearingDurationBenefitIssueCodes(
                 caseData,
@@ -307,8 +127,6 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
             + "getHearingDurationBenefitIssueCodes returns null")
         @Test
         void getHearingDurationBenefitIssueCodesNotPaper() {
-
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
 
             List<CcdValue<OtherParty>> otherParties = List.of(new CcdValue<>(
                 OtherParty.builder()
@@ -354,9 +172,15 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
             ))
                 .willReturn(DURATION_FACE_TO_FACE);
 
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-            SscsCaseData caseData = wantsToAttendCaseBuilder("Yes");
+            SscsCaseData caseData = SscsCaseData.builder()
+                .benefitCode(BENEFIT_CODE)
+                .issueCode(ISSUE_CODE)
+                .appeal(Appeal.builder()
+                    .hearingOptions(HearingOptions.builder()
+                        .wantsToAttend("Yes")
+                        .build())
+                    .build())
+                .build();
 
             Integer result = HearingsDurationMapping.getHearingDurationBenefitIssueCodes(
                 caseData,
@@ -378,8 +202,6 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
                 any()
             ))
                 .willReturn(DURATION_INTERPRETER);
-
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
 
             SscsCaseData caseData = SscsCaseData.builder()
                 .benefitCode(BENEFIT_CODE)
@@ -404,8 +226,6 @@ class HearingsDurationMappingTest extends HearingsMappingBase {
             + "getHearingDurationBenefitIssueCodes return the correct paper durations")
         @Test
         void getHearingDurationBenefitIssueCodesNotAttendNotPaper() {
-
-            given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
 
             SscsCaseData caseData = SscsCaseData.builder()
                 .benefitCode(BENEFIT_CODE)
